optimized some code
