+++ src/components/resource/resource.styles.css
@@ -3,12 +3,9 @@
 }
 
 .layer{
-        height: 600px;
-        width: 1450px;
-        margin-left: -80px;
-        margin-top: -80px;
+        height: 100%;
+        width: 98%;
         position: absolute;
         z-index: 10;
-        opacity: 0;
-        background-color: black
+        opacity: 0
 }
