--- src/components/review/review.styles.css
@@ -1,22 +1,0 @@
-.display-box{
-    height: 0px;
-    width: 0px;
-    margin-left: 35%;
-    margin-top: 10%;
-    border: solid black 1px;
-    transition-duration: .6s;
-    border-radius: 0%;
-    opacity: 0;
-    position: absolute;
-    font-size: 0px;
-}
-
-.display-box.active{
-    height: 350px;
-    width: 300px;
-    border-radius: 15%;
-    opacity: 1;
-    font-size: 17px;
-    background-color: white;
-    z-index: 20
-}
